requirements.txt
celery==3.0.8
sphinx
twisted
pyyaml
